0.8.20.17: 	suppressed a VERIFY-BACKTRACE test, because it seems to be 		broken independent of the haiblefixes I'm merging 	merged various fixes from Bruno Haible sbcl-devel 2005-03-10. 		(Note that some can't be properly exercised w/out a 		Mac or CLISP, but since they were self-evidently broken 		before (e.g., #+DARWIN or SUBTYPEP where only #!+DARWIN                 or SB!XC:SUBTYPEP makes sense), I merged them on the 		theory that it can't be making things fundamentally 		worse.:-) 	(+ unrelated .cvsignore tweaks to reduce general CVS nagging)

+++ src/runtime/interr.c
@@ -136,13 +136,13 @@ describe_internal_error(os_context_t *context)
 #ifdef sc_WordPointerReg
 	case sc_WordPointerReg:
 #endif
-	    printf("\t0x%08lx\n", *os_context_register_addr(context, offset));
+	    printf("\t0x%08lx\n", (unsigned long) *os_context_register_addr(context, offset));
 	    break;
 	case sc_SignedReg:
-	    printf("\t%ld\n", *os_context_register_addr(context, offset));
+	    printf("\t%ld\n", (long) *os_context_register_addr(context, offset));
 	    break;
 	case sc_UnsignedReg:
-	    printf("\t%lu\n", *os_context_register_addr(context, offset));
+	    printf("\t%lu\n", (unsigned long) *os_context_register_addr(context, offset));
 	    break;
 #ifdef sc_SingleFloatReg
 	case sc_SingleFloatReg:
